requirements.txt
asgiref==3.3.1
awesomeversion==21.2.3
beautifulsoup4
browserstack-local
cryptography==3.3.2
celery==5.1.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
django-rest-auth[with_social]
djangorestframework
django-filter
django-datatables-view
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crontab
django_cron
gunicorn==20.1.0
gevent
jsbeautifier
markdown
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
Pillow
pyimgur
redis==4.1.4
requests==2.9.1
six==1.15.0
selenium
Unidecode
undetected-chromedriver
webdriver-manager
whitenoise